Simplify away the `nodeEncoderChooser` struct altogether
